requirement.txt
fastapi
pydantic
pyodbc
python-dotenv
langchain
langchain-openai
